moved tests into test/unit
