+++ .env.example
@@ -1,11 +1,10 @@
 This is a example, this is not the actual login:
 
-E2E_EMAIL=your-test-user@example.com
-E2E_PASSWORD=replace-me
-
 My example:
 
 E2E_EMAIL=test@stud.noroff.no
 E2E_PASSWORD=testtest123
+
+
